Strip out server stuff and clean up tests (#1)
Stripped out everything not related to the core game logic so that we can get down to a basic minimal working demo of card interactions using unit tests.

Cleaned up the packages situation and reset everything with up-to-date packages.

Added a devcontainer.

Tests seem to be working apart from some failures that look to be genuine card logic issues.

Card data import is working! Added an npm command for it.