Refactor Card.ts into appropriate subclasses using mixins (#21)
Major refactor of Card.ts as it was a monolithic property class that held potential properties and methods for all card types together, making it easy to make mistakes about which properties / methods were available to use in which situation.

Since there is not a clean inheritance tree for card types, made as much of a tree as possible and then used mixins to add properties at multiple points in the tree. Added new methods for determining what type a card is so that its correct usage can be validated easily, and added helpers for creating and easily casting cards to the right type.