Added "gain ability" support and Vambrace Grappleshot as example (#25)
Re-enabled the `GainAbility` stuff from L5R to support dynamically
gained abilities. This is particularly important for upgrades as most of
them grant a triggered ability to the attached card.

Added Vambrace Grappleshot as an example. The same tools will work for
units that grant dynamic abilities.

Also did some cleanup of the upgrade test infrastructure to fix a bug
that was causing sporadic failures.