Implement events, add Daring Raid and Repair (#26)
Implemented the code for Event card types. Added Daring Raid and Repair
to test the system. Need to add Restock to test that the event
sequencing happens right, but waiting on other changes to merge first.

Also fixed a bug with upgrade aspect costs and added a test for it, as
well as some general cleanup.