Finish Base implementation, add Tarkintown (#28)
Finalized the infrastructure for adding Base cards. Added Tarkintown as
an example. Wanted to test that the epic action limit is preserved
across turns, so did a lot of cleanup + debugging of the non-action
phases to enable the option to move to a second action phase during the
test.

This ended up meaning a lot of cleanup of multiple parts of the code,
but now the regroup and setup phases are known to be working correctly
(though claim initiative is not currently tested).

---------

Co-authored-by: Anthony Bronico <adbronico92@gmail.com>