Add support for token upgrades (#30)
Added infrastructure to support token upgrades. Shield and Experience
tokens are now generated by the game using the GiveExperienceSystem or
GiveShieldSystem. The test infrastructure was updated to not try and
move the token cards into the deck, they are instead generated live as
the zones are populated.

As part of shield, added a new window for "replacement effects" which
modify the standard effect of a triggered ability / event.

Realized that the way we had units grouped in the enums didn't quite
work with tokens, separated out the "non-token, non-leader" types as
"basic" types and added more wildcard filters to cover all cases.

Needed a few extra event cards in place to test that all of the token
behaviors are expected, added those as well. Also fixed some bugs in the
test setup code.