Add support for leaders (#34)
Added the full functionality of leaders. Simplified and cleaned up the
exhaust and pay resource systems a little bit. Updated the documentation
to reflect the process for leaders and some potential pitfalls.

Added Krennic and Tarkin leaders to proof the system. Added a general
test file for leader behaviors and some new test infra to support leader
testing.

Some other general improvement:
- Some bug fixes and improvements to the test infra
- Added type guards on the `isX()` methods on `Card.ts` so that they can
be used to remove casts in certain situations