Add state watchers and a few example cards (#35)
Added the concept of state watchers to address the needs of cards that
refer back to events that have happened previously in the phase. For
now, not making a distinction between "this round" and "this phase"
since there is currently no game situation in which the distinction is
important.

Added the Luke leader, Medal Ceremony, and Vanguard Ace as examples.
Also did some light improvement and bug fixes to the test infra.