changed Tie Advanced, added Wing Leader and attempted to write the Wing Leader unit test. (#38)
I noticed that the template I was using for Wing leader called for two
constants that it did not use and removed those. Then I copied the Tie
Advance to implement the wing leader. I then used the Tie Advance unit
test to create a unit test for Wing leader.

---------

Co-authored-by: teadams5 <teadams5@asu.edu>