Fix "when defeated" trigger issues, add Vanguard Infantry (#44)
We discovered a bug causing when-defeated triggers to completely fail.
The reason was that triggers were being activated after events were
resolved, so if a card left the arena during the event resolution window
(e.g. due to defeat resolving) then its abilities would be unregistered
and the defeat ability would never fire.

As a (possibly temporary) mitigation, we changed the resolution to work
as follows:

1. Triggers are activated before event resolution, all triggered
abilities are saved
2. Resolve event(s)
3. Activate triggers _again_ to catch any triggers such as when-played
that would've been missed at step 1. Duplicate triggers are ignored.
4. Check if any triggering events were cancelled; if so, remove the
relevant triggered abilities from the list to resolve
5. Resolve all remaining triggered abilities

If the ability is a "then" ability (i.e. triggers are being held until
the parent ability finishes resolving), we don't active any triggers but
instead add them to the parent triggered ability window.

Added Vangard Infantry as a proof card to confirm that everything is
working.

TODO: we need to investigate whether it would be better for cards to
have their abilities online while not in the arena, to simplify cases
like this. It's also needed for exploit and possibly ambush.