Fix triggering requirements for triggered and action abilities (#45)
L5R has different rules than SWU for allowing triggered abilities and
actions to active. Made some updates to correct those requirements so
our cards can function correctly.

**Triggered abilities:** this was pretty straightforward, there was
already an "optional" flag that we could use in cases like Death Trooper
for allowing a targeted to fizzle. It just wasn't working as expected
for some reason, needed some additional implementation.

**Action abilities:** went through and changed how `meetsRequirements`
works so that if the ability is an action and has a cost that impacts
game state, it's automatically allowed to activate.

Needed to add some changes to the way card selectors work to satisfy the
above. Added tests for some units that need the above and confirmed that
they pass.

Also did some cleanup to prevent the use of multiple types of system
specifiers (initiateAttack, immediateEffect, targetResolver(s)) on the
same ability since the interactions between these are non-obvious and
error-prone.