Added Headhunting and the SequentialSystem (#46)
Added a new meta-system back from jigoku, the SequentialSystem. This is
used to chain a sequence of effects together with a shared triggering
window at the end (effectively the same as a set of "then" effects).

Demonstrated the use of it with the Headhunters card, along with the
SelectCardSystem for making card selections in a context where target
resolvers aren't allowed - in this case, inside the chain of systems
within SequentialSystem. Did some cleanup in there as well.