Add system for "if" conditions and some cards to test triggered and constant abilities (#51)
Rounded out the ConditionalSystem so that it could be used to support
most cases of the word "if" in card abilities (the primary exception
being "if you do", which will require special logic).

Implemented a few cards to demonstrate how to implement the word "if"
when used in:
- Normal triggered abilities
- Triggered abilities on attachments
- Constant abilities on attachments

Examples for other cases (particularly actions) will come in future PRs
but the core functionality should be the same for all.

Also did some improvements to engine components to resolve issues found
during testing, mainly in the TriggeredAbilityWindow. The fact that it
would error out in a situation where there are multiple triggers on the
same unit became a blocker, so for now we are just presenting every
ability by text name regardless of any additional context. This will
need to be improved eventually, but for the simple cases that come up in
tests it's fine for now.