Add a debugging guide (#50)
Added a debugging guide to help contributors get started with debugging
if they run into any issues. Also added a cheat sheet for unit tests
that currently just lists the current set of custom expectations that
are available.

In the process of writing the cheat sheet, realized one of the
expectations had a bug which I fixed.