Changes to trigger resolution to more closely match the rules. (#54)
Constructing an EventWindow now specifies whether or not triggers
triggered during that window should resolve right after those events. If
this is set to false, the EventWindow will add the triggers triggered by
its events to the TriggeredAbilityWindow belonging to its parent
EventWindow. This structure implements the game rules concept of nested
actions and triggers.