Add `simultaneous` meta-system and The Force is With Me (#58)
Add the second meta-system, `SimultaneousSystem`. This is for the (more
common) case where multiple effects resolve in the same triggering
window. Add The Force is With Me as a test card.

Since L5R had multiple ways of handling abilities with multiple effects
which did not clearly state the differences between them, we're
consolidating down to "sequential" and "simultaneous" and we'll expand
as needed. Thus we're removing the option to provide an array of
GameSystems as an immediate effect, instead forcing people to use
`AbilityHelper.immediateEffects.simultaneous()` to make the intent
clear.

Updated the documentation to include how to use the new systems, as well
as backfilling some documentation for other recent PRs.

Fixed a bug in the `.toHaveExactUpgradeNames()` test expectation.