Add Snoke and logic for defeat due to HP effects (#73)
Added the engine infrastructure to handle defeat due to HP modifying
effects.

Add the concept of "subwindows" to EventWindow. A subwindow is a new
window that resolves immediately after the current window's events
resolve, before triggered abilities resolve. This is needed to account
for the timing of defeat due to combat, damage, etc.

Added Snoke and several tests that his ability works and that relevant
triggered abilities happen in the right time window. There are more
tests that need to be added for more advanced interactions involving
multiple layers of resolving constant abilities / nested triggered
abilities, but these will be reserved for a dedicated future PR.

Also added Dodonna and some basic tests that his ability interacts
correctly with defeat. More tests are needed but are blocked by
implementing units that would be needed for the test.

Added a few additional units for testing purposes, keeping it to the
simplest possible ones. More complicated cases reserved for the future
timing windows test PR.