Add support for "optional" parameter in targetResolvers (#103)
Re-added support for `optional: true` in target resolvers, in which case
they will present the option "Choose no target." Aligned the behavior of
`selectCard()` to this behavior as well.

Clarified the language of passing at the "choose attack target" stage of
an attack to say "Pass attack," though this is likely temporary while we
decide on a final flow for this. Added more specific test helpers for
all of these new buttons.