Claim Initiative, Test Typings, and Baze Malbus Implementation (#111)
## Initiative claiming functional
claiming initiative now properly flags the player as "passed", simply
had it call pass after it finishes, and the button to claim initiative
is removed for the rest of the phase. Added an event window on claiming
initiative, but nothing specifically tests it yet.

## server and test build changes
I reorganized the build process somewhat. The server build is now output
to ./build/server by default, and test builds to ./build/test. server
and test build independently of each other but test is still aware of
server. JS/TS files under ./tests can now reference '../server' instead
of '../build' as the folders are in the same relative position, which
provides better typing support and more intuitive require(...) paths.

There is a now a npm script `npm run test-fast` that can be used the
same as `npm test`. This will only rebuild the test folder rather than
both the server and test folders, unless it detects that the server
hasn't been built yet, at which point it just does a full rebuild. Might
make re-running a single test file a little faster but shouldn't be used
if you're also changing server code.

`npm test` can build both the server and test folders in parallel now.

Added a sub-folder for custom npm scripts, ./scripts. These just are for
doing npm commands with a little more complexity. Not absolutely
required, but felt it was handy to have.

the new eslint configs messed with building TS files in the test folder,
so I kinda just hit it with a hammer until it stopped complaining. Will
probably need a refactor from someone who has a better understanding of
it.

## Test changes
hasInitiative can now be flagged in the setupTest options player object.
This will properly setup the given player has having the initiative at
the start of the game.

While digging into that, I also changed a few arrays calls, some
functions were sorting the array's in-place which was throwing off some
testing. Changed all calls to use .concat() first to create an array
copy to then sort, and separately fixed a check or two to properly sort.

### General typing improvements
I've added an IntegrationsHelper.d.ts file. Trying to convert the JS
file to TS wasn't worth the hassle due to heavy usage of `this` and
other side-effects caused by integration. Custom Matchers have been
added so they should now show up in intellisense for tests when using
`exact()`. There are typings for the options object and the some of the
game objects, but require using a new test context object to get
typing/intellisense for.

### limited TS support and test context
With the test folder structure changes, we can now create TS specs.
However, it's not feasible to make `this` type safe as-is (better to
just rewrite in vitest). As a workaround it's now possible to pass a
contextRef in with the integration function call ala
`integration(function(contextRef) { ... })`. This has a `context` object
which contains the exact same fields that are accessible via `this` with
some additional typings, using reference sorcery.

Instead of calling `this.setupTest({ ... })`, you can call
`contextRef.setupTest({ ... })`. In a given `it(...)` block, you can
destructure the contextRef to get the context specific to that test,
`const { context } = contextRef`, and use context as if you were using
`this`. This gives much stronger typing and intellisense support.

contextRef is entirely optional, and `this` inside of tests functions
exactly the same as before.

NOTE: This doesn't take into account any cards defined under the player
objects, so the context has a general purpose "any" accessor. No type
safety but at the very least it's no worse than before.

See the specs for examples.

## Oh right, Implemented Baze Malbus, Temple Guardian
Dude's implemented.