Fix windows so that "gain ambush" and "gain shielded" work correctly (#120)
We had a bug where "gain ambush" and "gain shielded" would not work
because units register those abilities when they move into the arena,
but the effects that give them abilities resolve after they get to the
arena.

Made a few changes to address this. The first is refactoring the event
window so that we emit events again after a `resolveGameState` call so
that any abilities gained during the event window resolution can
successfully trigger.

The other change is refactoring how we handle triggered abilities for
keywords. We previously would register keywords at a time point before
the event actually resolves (e.g., registering when-played abilities
when the unit enters the arena). However, it is possible that gain or
lose ability effects can happen in between those two time points, so we
need to evaluate whether the unit has the keyword exactly at resolution
time.

To facilitate this, refactored how the triggering for keyword abilities
happens - instead of pre-registering the abilities at an earlier point,
we now check every unit that is played or declares an attack to
determine if it has any relevant keywords when the attack / play is
resolved, once it is past the point that any effects could add / remove
keywords. This guarantees that we will always correctly determine the
relevant set of keywords for the card.