Basic server mvp (#119)
This should be the bare minimum needed to get a game node up and running
locally. `server/gamenode` contains the GameServer class as well as a
default setup for creating a single game with dummy data. Once built you
can run `node build/server/gamenode` to run the server which
automatically starts and creates the default game.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>