Add C-3PO and list select prompt (#121)
Implemented the C-3PO card. As part of that, added a target resolver and
prompt for getting the player to choose from a dropdown list of options.

This hit a lot of bugs and weak areas of the code for some reason, had
to do a lot of minor bug fixes and cleanup to get it working.