Tracking damage / defeat source for "attacks and defeats" and similar (#124)
Several abilities will ask what the cause of a unit being damaged or
defeated is, the most frequent being "when [X] attacks and defeats a
unit." To implement this we need a little bit of provenance tracking on
certain types of events (damage and defeat) so we can identify the
source.

Added some basic mechanisms for this, implemented Rukh and Mace Windu to
test the new features.

Had to leave a partial gap because the rules for attributing the player
that caused a defeat aren't fully fleshed out when it comes to hp
modifying effects like Snoke and Dodonna. FFG has promised this will be
included in comp rules 3.0 and it only matters for Bravado right now so
just added TODOs to cover this for the moment.