Add unique rules (#126)
Implemented the rules for enforcing unique cards in play. Added tests
for the edge cases of Yularen and Snoke to confirm that they work
correctly.

Yularen worked out of the box. Snoke required adding some additional
logic to disable ongoing effects when a card has a pending defeat. Also,
refactored the "pending defeat" code to be clearer and operate more
consistently.