Safer handling of triggers in event window (#144)
We've had a couple bugs due to the fact that sometimes a trigger window
is set to pass its triggers to the parent but the parent is not a window
that should have triggers and therefore they resolve at the wrong time
or not at all.

In order to catch problems like this early, changed the trigger mode
from a boolean to an enum with three options: resolves triggers, passes
triggers to parent, should not have triggers. The last option is for
"meta-events" which are events that exist in our framework as utilities
to support proper game function instead of being actual SWU rules events
which abilities could respond to.

In making the above changes I caught a similar issue with the regroup
phase where triggered abilities would not be resolving correctly (we
just didn't have any abilities triggering on that event yet).

Additionally, updated GameSystem so that it forces every child class to
explicitly declare an event name as some systems were not doing so and
were emitting "Unnamed" by default. Removed the "Unnamed" event
completely and added a category of event names for meta-events that
should not be triggering any abilities (MetaEventName).