Support for "if you do (not)" clauses (#154)
Added support for "do [X], if you do (not) then do [Y]" clauses via the
`ifYouDo: ` property which works similarly to the `then :` property. It
will key off of the effect in the ability to see if it resolves or not,
following the rules of SWU resolution (must change game state or be
replaced by an effect that does). If there are multiple effects in the
ability, it will key off of the resolution of the last one.

Had to heavily refactor to add a better API for managing events and
checking their resolved / resolvable status for this. Added a
`mustChangeGameState` parameter to many of the canAffect /
hasLegalTarget flows to enforce checking whether events have resolved or
not in a way that changes game state so that "if you do" conditions will
evaluate correctly.

Added the following cards for validation:
- Boba Fett, Daimyo (updated to use if you do instead of cost)
- Enforced Loyalty
- Toro Calican
- Lom Pyke

Discovered a bug with replacement effect ordering in the Lom Pyke tests.
Applied a partial fix but a full one will have to wait for a dedicated
PR to build the real replacement effect window. That said, the condition
that it applies in is extremely rare so it shouldn't be an issue until
then.