Refactor Zone and Player enums (#212)
In preparation for adding the Zone class in a soon PR, renamed
"Location" to "Zone" throughout the repo. Since this is massively noisy,
did some other updates to enums that were needed but also ended up being
quite loud.

The major onus for this is that the Zone class will have an enum value
ZoneName that corresponds to the current Location enum. Using the term
"Zone" is more in line with the rules overall, so it was a good
opportunity to clean that up. Notable changes in this PR:

- Location / WildcardLocation enums are now ZoneName / WildcardZoneName
- LocationFilter is now ZoneFilter
- Converted the string representations of the enums to be camel case in
line with all other enum strings: e.g., `'ground arena'` ->
`'groundArena'`
- Split the RelativePlayer enum into RelativePlayer (Self, Opponent) and
WildcardRelativePlayer (Any) joined in the type RelativePlayerFilter
- Changed the Player.countSpendableResources() and
Player.countExhaustedResources() to Player.readyResourceCount and
Player.exhaustedResourceCount