Zone class for managing card locations (#215)
Added a tree of Zone classes for managing the different zones that cards
can occupy in the game. Overhauled to migrate everything onto the new
zone framework.

Zones have a lot of useful features, most notably a `getCards()` method
that accepts a flexible filter object and will return cards that match
the filter based on trait, aspect, type, etc.

Will do a future PR to migrate more of the existing code onto using this
new API, for this PR just refactored under the hood and left the
existing ones in place.