Target resolver bug fixes and improvements (#259)
Encountered a bug with target resolvers where if the resolver was marked
optional (not the ability) and there was only a single target, it would
still auto-select the target.

In the process of fixing that, ended up doing some overall refactoring
of the interfaces to make things cleaner and more consistent. Fixed
things so that the prompt explaining the hidden zone rule now appears
when expected, and also removed the option to have a target resolver be
declared optional unless it's one of many.