Bug fix for gained ability from lasting effect (#265)
In another branch we hit an issue where a gained ability couldn't
correctly generate a name for itself when created from an attack lasting
effect due to the lack of a name on the effect (or on the originating
ability). Added a way to declare when the ability is coming from a
lasting effect. Tests pass in the feature branch.