Add check for open prompt at end of test cases (#320)
A common hangup in testing, especially for new folks, is not realizing
that a prompt is open and waiting for a click. To help people catch
this, added a check at the end of every test case which will throw an
error if either player has an open prompt (so long as they end in the
action phase).

This functionality can be disabled by setting
`context.allowOpenPromptAtEndOfTest = true`. Updated all affected
existing test cases to either not end with an open prompt or to use the
above setting to disable the check.