Disable autoSingleTarget by default in tests (#324)
Since lots of people were getting tripped by the autoSingleTarget
feature causing automatic selection of one available target during
tests, added a feature to control the setting in the test setup and
turned it off by default. For existing tests that depend on the feature,
manually set the flag to true with a comment warning not to do the same
in new code.

Updated a couple of the older tests to have the flag off just to confirm
that everything works right.