Don't do open prompt check if test failed (#333)
We had an issue where the open prompt check at the end of a test would
trigger even if the test failed due to known reasons (exception or
expectation failure). This creates confusion so I added a check to keep
it from happening.

Unfortunately the mechanism for this doesn't work in parallel mode, so
in that case I added an error message indicating that we need to run in
serial for more details.