Add Multi-Troop Transport and add Exploit parsing (#345)
The engine was choking on cards with the Exploit keyword since it didn't
have an entry for it to parse. Added the logic for that and tested by
adding Multi-Troop Transport.

Exploit functionality is not present at all but the keyword can at least
be parsed successfully and then ignored.