Add support for gaining Saboteur for attack duration (#355)
There was an issue preventing us from implementing cards that allow a
unit to attack and gain Saboteur for that attack. Specifically, the
"gain Saboteur" effect would not be applied until after targeting,
meaning that it wouldn't have any impact on sentinel.

Hardcoded support for this into the AttackStepsSystem and implemented
Breaking In to test it. Also did some general cleanup of the
CardLastingEffectSystem and AttackStepsSystem.