Fix delayed effect targeting bug (#384)
We hit an issue with the Sneak Attack branch where a delayed effect
weren't correctly having its `target` property set but the error was
extremely unclear. In addition, the target resolution wasn't being
handled correctly if `selectCard` is used. Added better error checking
and a clearer message, and fixed the target resolution.