Bossk leader (#405)
Refactored the bounty resolution process to create an explicit
"onBountyCollected" event, and made it possible to store the properties
of the bounty keyword in the event. Added the CollectBountySystem which
the Bossk ability uses to re-trigger the bounty ability.

Added a big test suite to cover the various types of bounties and ensure
that all can double-resolve correctly. Added some specific detailed
tests for the Bounty Hunter's Quarry case, including an open TODO on one
case while we wait for clarification on some rules.