Implementing IG-11, I Cannot Be Captured (#459)
![image](https://github.com/user-attachments/assets/47dda194-39ca-40a2-ae86-6e8a536fd484)


This required a small refactor in the `ReplacementEffectSystem.ts` and
`SimultaneousGameSystem.ts` since they didn't support multiple effects
replacing a single effect. I also added unit tests to cover multiple
defeat effects to make sure that they are all getting resolved correctly
with the refactor. The only issue I'm not sure on is the
`GameEvent.setReplacementEvent`. Currently the event will just be
whichever was last. So far this doesn't seem to be negatively affecting
anything. If at some point we do need to track all of the events that
did the replacing, the `replacementEvent` could be turned into an array.