Fix in-play id issue for hidden zones (#511)
We had an issue where the "in play id" which identifies the current "in
play copy" of a card was not being updated correctly. When a card leaves
a visible non-arena zone (e.g. discard) and goes to a hidden zone (e.g.
hand), it should become a new copy of the card even if it then returns
immediately to discard.

This was causing issues in the Admiral Trench case since he was seeing
units that were defeated, returned to hand, and discarded again as still
the same copy. This change fixes the bug in that branch.